refactor
This is just to expose a potential "user pain" area when it comes to invoking Actions.
The issue being that the exception message provided to the user if they are late joining and have mistakenly forgotten to remove a previously used NetworkBehaviour's registered delegate from the list of registered delegate handles.

The end result under this particular scenario is that late joining players will not get registered with the other players because an exception is thrown when trying to invoke a method on a disposed NetworkBehaviour.